HW02.2 - refactoring Meal and MealTo
