+++ src/main/java/ru/javawebinar/topjava/to/MealTo.java
@@ -1,4 +1,4 @@
-package ru.javawebinar.topjava.model;
+package ru.javawebinar.topjava.to;
 
 import java.time.LocalDateTime;
